More robust TCs

+++ Content.Server/Traitor/Uplink/Telecrystal/TelecrystalSystem.cs
@@ -1,9 +1,11 @@
 using Content.Server.Traitor.Uplink.Account;
 using Content.Server.Traitor.Uplink.Components;
 using Content.Shared.Interaction;
+using Content.Shared.Popups;
 using Content.Shared.Stacks;
 using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
+using Robust.Shared.Localization;
 using System;
 
 namespace Content.Server.Traitor.Uplink.Telecrystal
@@ -21,8 +23,15 @@ namespace Content.Server.Traitor.Uplink.Telecrystal
 
         private void OnAfterInteract(EntityUid uid, TelecrystalComponent component, AfterInteractEvent args)
         {
+            if (args.Handled)
+                return;
+
             if (args.Target == null || !EntityManager.TryGetComponent(args.Target.Uid, out UplinkComponent? uplink))
                 return;
+
+            // TODO: when uplink will have some auth logic (like PDA ringtone code)
+            // check if uplink open before adding TC
+            // No metagaming by using this on every PDA around just to see if it gets used up.
 
             var acc = uplink.UplinkAccount;
             if (acc == null)
@@ -31,9 +40,16 @@ namespace Content.Server.Traitor.Uplink.Telecrystal
             EntityManager.TryGetComponent(uid, out SharedStackComponent? stack);
 
             var tcCount = stack != null ? stack.Count : 1;
-            _accounts.AddToBalance(acc, tcCount);
+            if (!_accounts.AddToBalance(acc, tcCount))
+                return;
 
             EntityManager.DeleteEntity(uid);
+
+            var msg = Loc.GetString("telecrystal-component-sucs-inserted",
+                ("source", args.Used), ("target", args.Target));
+            args.User.PopupMessage(args.User, msg);
+
+            args.Handled = true;
         }
     }
 }
